% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genr_clean.R
\name{clean_spss_bs}
\alias{clean_spss_bs}
\title{Transform all "SPSS labelled" to r base types}
\usage{
clean_spss_bs(df, use_labels_as_levels = TRUE)
}
\arguments{
\item{df}{A dataframe (optionally labelled)}

\item{use_labels_as_levels}{For factors.}
}
\description{
Takes a dataframe and transforms all variables with class \code{haven_labelled}
to factors, numeric or character vectors. Preserved variable labels if there are any.
}
