% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genr_clean.R
\name{factorize}
\alias{factorize}
\title{Transform all factors-looking things to factors}
\usage{
factorize(df, min_levels = 10)
}
\arguments{
\item{df}{A dataframe (optionally labelled)}

\item{min_levels}{Minimum number of unique values, above which the variable
is not treated as a factor. Default = 10.}
}
\description{
Takes a dataframe and transforms all variables with less than \code{min_levels}
unique values to factors. Preserved variable labels if there are any.
}
