% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{gr_read}
\alias{gr_read}
\title{Read and clean .sav files}
\usage{
gr_read(file, datapath = NULL, summary = FALSE)
}
\arguments{
\item{file}{: file name (with or without extension; with or without full path
specification).}

\item{datapath}{: (NULL) allows to specify the path where the file is located.}

\item{summary}{: (FALSE) print the summary of all variables in the dataset.}
}
\value{
An R data.frame with variables as columns and observations as rows.
}
\description{
The function allows to quickly load and clean a .sav file into an R dataframe.
It requires a \code{file} name that can be with or without extension (i.e., '.sav')
specified. The path to the file can be either included in the file name, or
specified via the \code{datapath} argument. If no (existing) path is provided, a
navigation window will be prompted and the user can decide whether to save the
path for later loading jobs.

If the file is correctly identifies this is loaded into an R data.frame which
preserve value labels and substitutes 999 and 888 values with NA where
appropriate.
}
\examples{
data1 <- gr_read('filename')
data1 <- gr_read('filename.sav')
data1 <- gr_read('filename', datapath='path/to/file/')
data1 <- gr_read('filename', summary=TRUE)

}
