% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genr_clean.R
\name{spss_labelled_to_factor}
\alias{spss_labelled_to_factor}
\title{Transform all "SPSS labelled" to factors}
\usage{
spss_labelled_to_factor(df)
}
\arguments{
\item{df}{A dataframe (optionally labelled)}

\item{min_levels}{Minimum number of unique values, above which the variable
is not treated as a factor. Default = 10.}
}
\description{
Takes a dataframe and transforms all variables with class \code{haven_labelled}
to factors. Preserved variable labels if there are any.
}
