% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genr_info.R
\name{data_overview}
\alias{data_overview}
\title{Print an overview of the dataset}
\usage{
data_overview(df)
}
\arguments{
\item{df}{A dataframe (optionally labelled)}
}
\description{
Prints a table to console with info about all the column names, labels and
types (e.g., factor, numeric...).
}
